backend/requirements.txt
SQLAlchemy==2.0.20
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-RESTful==0.3.10
Flask-Limiter==3.5.0
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.3
python-dotenv==1.0.0
Pillow==10.0.0
marshmallow==3.20.1
